Silence warnings if endless-image-device is not mounted
If you try to mount an already-mounted ext4 filesystem at a new
location, it succeeds (you get a bind mount). But this doesn't work for
ntfs3g; so we try to umount the C: drive before continuing. If it's not
mounted, that's fine, we just don't want any scary warnings.